Restructure campground routes using express router
