+++ app.js
@@ -9,6 +9,8 @@ const ExpressError = require("./utils/ExpressError");
 const Campground = require("./models/campground");
 const Review = require("./models/review");
 
+const campgroundRoutes = require("./routes/campgrounds");
+
 mongoose.connect("mongodb://127.0.0.1:27017/yelp-camp");
 const db = mongoose.connection;
 db.on("error", console.error.bind(console, "Connection error!"));
@@ -22,16 +24,6 @@ app.set("views", path.join(__dirname, "views"));
 app.use(express.urlencoded( { extended: true}));
 app.use(methodOverride("_method"));
 
-const validateCampground = (req, res, next) => {
-    const {error} = campgroundSchema.validate(req.body); // destructure error from result object
-    if(error) {
-        const message = error.details.map(el => el.message).join(","); // error.details is array of objects so we map 
-        throw new ExpressError(message, 400);                         // over them and make a string of them
-    } else {
-        next();
-    }
-}
-
 const validateReview = (req, res, next) => {
     const { error } = reviewSchema.validate(req.body);
     if(error) {
@@ -42,42 +34,7 @@ const validateReview = (req, res, next) => {
     }
 }
 
-app.get("/campgrounds", async (req, res, next) => {
-    const campgrounds = await Campground.find({});
-    res.render("campgrounds/index", {campgrounds});
-});
-
-app.get("/campgrounds/new", (req, res) => {
-    res.render("campgrounds/new");
-});
-
-app.get("/campgrounds/:id/edit", catchAsync(async(req, res, next) => {
-    const campground = await Campground.findById(req.params.id);
-    res.render("campgrounds/edit", {campground});
-}));
-
-app.put("/campgrounds/:id", validateCampground, catchAsync(async (req, res, next) => {
-    const {id} = req.params;
-    await Campground.findByIdAndUpdate(id, {...req.body.campground}, {new: true});
-    res.redirect(`/campgrounds/${req.params.id}`);
-}));
-
-app.post("/campgrounds", validateCampground, catchAsync(async (req, res, next) => { 
-    const campground = new Campground(req.body.campground);
-    await campground.save();
-    res.redirect(`/campgrounds/${campground.id}`);
-}));
-
-app.get("/campgrounds/:id", catchAsync(async (req, res, next) => {
-    const campground = await Campground.findById(req.params.id).populate("reviews");
-    res.render("campgrounds/show", {campground});
-}));
-
-app.delete("/campgrounds/:id", catchAsync(async (req, res, next) => {
-    const {id} = req.params;
-    await Campground.findByIdAndDelete(id);
-    res.redirect("/campgrounds");
-}));
+app.use("/campgrounds", campgroundRoutes);
 
 app.delete("/campgrounds/:id/reviews/:reviewId", catchAsync(async(req, res) => {
     const { id, reviewId} = req.params;
